+++ kinto_attachment/tests/test_plugin_setup.py
@@ -16,9 +16,3 @@ class HelloViewTest(BaseWebTestLocal, unittest.TestCase):
             "base_url": "https://files.server.com/root/"
         }
         self.assertEqual(expected, capabilities["attachments"])
-
-    def test_public_url_is_provided_in_public_settings(self):
-        resp = self.app.get("/")
-        settings = resp.json["settings"]
-        self.assertEqual(settings["attachment.base_url"],
-                         "https://cdn.firefox.net/")